requirements.txt
pandas
numpy
scikit-learn
rouge-score
sentence-transformers
transformers
openai
tqdm
matplotlib
seaborn
typer
python-dotenv
pyyaml
